+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.1] - 2025-05-14
+
 ### Added
 
 - When a route is used which is about to expire, we now send a route request to
@@ -603,4 +605,5 @@ This is a breaking change, check the main README file for update info.
 [0.5.6]: https://github.com/threefoldtech/mycelium/compare/v0.5.5...v0.5.6
 [0.5.7]: https://github.com/threefoldtech/mycelium/compare/v0.5.6...v0.5.7
 [0.6.0]: https://github.com/threefoldtech/mycelium/compare/v0.5.7...v0.6.0
-[unreleased]: https://github.com/threefoldtech/mycelium/compare/v0.6.0...HEAD
+[0.6.1]: https://github.com/threefoldtech/mycelium/compare/v0.6.0...v0.6.1
+[unreleased]: https://github.com/threefoldtech/mycelium/compare/v0.6.1...HEAD